Feat : 주문 컨트롤러 수정

+++ src/main/java/com/konkuk/bit/bitpay/order/dto/OrderCreateDto.java
@@ -13,6 +13,6 @@ public class OrderCreateDto {
     /*
         menuNumber, menuQuantity
      */
-    private Map<Integer,Integer> orderDetail;
+    private Map<Long,Integer> orderDetail;
     private Integer tableNumber;
 }
